==> Model/2)Preprocessing_QA.ipynb <==
# Preprocessing

## Library

# !git clone https://github.com/ssut/py-hanspell.git
# !pip install kss

# 수정한 hanspell 불러오기
from py_hanspell.hanspell import spell_checker
from py_hanspell.hanspell.constants import CheckResult

import numpy as np
import pandas as pd
import kss
import re
from tqdm.notebook import tqdm
import os

import warnings
warnings.filterwarnings("ignore")

# pd.set_option('display.max_columns', None)
# pd.set_option('display.max_rows', None)

## Path

base_path = '../Data/Base/'
save_path = '../Data/QnA/'

## Load Data

# qna 파일 불러오기
df = pd.read_csv(base_path+'qna_table.csv')

df[:1]

## 비밀글 & 답변 없음 제거

# 전체 개수
print(df.shape)

# 비밀글 제거 > 즉, 둘 중 하나라도 비밀글이면 drop
df_remove = df[(df['question']!='비밀글입니다.')&(df['answer']!='비밀글입니다.')]

# 제거한 후 개수
print(df_remove.shape)

# 답변 없음 제거
df_remove_ = df_remove[df_remove['answer']!='None']

# 제거한 후 개수
print(df_remove_.shape)

df_remove_[:1]

## 내용 중복 제거

print('질문 unique:', len(df_remove_['question'].unique()))
print('답변 unique:', len(df_remove_['answer'].unique()))

# 중복 제거
tmp_drop = df_remove_.drop_duplicates(['question','answer'], keep = 'first')
tmp_drop.reset_index(inplace=True, drop=True)

# 개수 확인하기
count = tmp_drop[['question','answer']].value_counts()
print('전체 개수: ', tmp_drop.shape)
print('중복이 없는 개수: ', len(count[count.values==1]))
tmp_drop[:1]

## 전처리

* 한글&숫자만 남기고 제거
* hanspell 맞춤법 검사 진행

df = tmp_drop

que = df['question'].to_numpy()
ans = df['answer'].to_numpy()

### 한글&숫자만 남기고 제거

def text_cleaning(texts): # 수정완료
    # 한국어를 제외한 글자를 제거하는 패턴.
    corpus = []
    for i in range(0, len(texts)):
        sentences = re.sub("[\{\}\[\]\/?.,;:|\)*~`!^\-_+<>@\#$%&\\\=\(\'\"]", " ", str(texts[i])) #특수문자를 제거하는 패턴
        sentences = re.sub("[^ㄱ-ㅎㅏ-ㅣ가-힣 |0-9|]", " ", sentences)
        sentences = re.compile('[|ㄱ-ㅎ|ㅏ-ㅣ]+').sub(' ', sentences) # 자음, 모음만 있는 경우

        sentences = re.sub(r"^\s+", '', sentences) #remove space from start
        sentences = re.sub(r'\s+$', '', sentences) #remove space from the end

        sentences = re.sub(r'\s+', ' ', sentences) #remove extra space

        corpus.append(sentences)
    
    return corpus

pre_que = text_cleaning(que)
pre_ans = text_cleaning(ans)

# df에 업데이트
df['question_clean'] = pre_que
df['answer_clean'] = pre_ans

print(len(pre_que), len(pre_ans))

### hanspell 맞춤법

# que
def hanspell_(texts):
    spell_check = []
    for line in tqdm(texts):
        spell = spell_checker.check(line).checked
        spell_check.append(spell)
    return spell_check

spell_check_question = hanspell_(pre_que)

# 확인
print(spell_check_question[:5])

# df에 업데이트
df['question_spellcheck'] = spell_check_question

# ans
def hanspell_(texts):
    spell_check = []
    for line in tqdm(texts):
        spell = spell_checker.check(line).checked
        spell_check.append(spell)
    return spell_check

spell_check_answer = hanspell_(pre_ans)

# 확인
print(spell_check_answer[:5])

# df에 업데이트
df['answer_spellcheck'] = spell_check_answer

## 내보내기

df.to_csv(save_path+'qna_table_hanspell.csv', index=False)
