==> Model/2)Preprocessing_Review.ipynb <==
# Preprocessing

## Library

import numpy as np
import pandas as pd
import re
from tqdm.notebook import tqdm
import os

from py_hanspell.hanspell import spell_checker
from py_hanspell.hanspell.constants import CheckResult

## Path

base_path = '../Data/Base/Review/'
save_path = '../Data/Review/Review_spellcheck/'

## 한글&숫자만 남기고 제거 / hanspell 맞춤법

def text_cleaning(texts): # 수정완료
# 한국어를 제외한 글자를 제거하는 패턴.
    corpus = []
    for i in range(0, len(texts)):
        reviews = re.sub("[\{\}\[\]\/?.,;:|\)*~`!^\-_+<>@\#$%&\\\=\(\'\"]", " ", str(texts[i])) #특수문자를 제거하는 패턴
        reviews = re.sub("[^ㄱ-ㅎㅏ-ㅣ가-힣 |0-9|]", " ", reviews)
        reviews = re.compile('[|ㄱ-ㅎ|ㅏ-ㅣ]+').sub(' ', reviews) # 자음, 모음만 있는 경우

        reviews = re.sub(r"^\s+", '', reviews) #remove space from start
        reviews = re.sub(r'\s+$', '', reviews) #remove space from the end

        reviews = re.sub(r'\s+', ' ', reviews) #remove extra space

        corpus.append(reviews)
    return corpus

def hanspell_(texts):
    spell_check = []
    for line in texts:
        spell = spell_checker.check(line).checked
        spell_check.append(spell)

    return spell_check

review_list = os.listdir(base_path)

# 실행
for review in tqdm(review_list):
    # 불러오기
    rev = np.load(review_path+'/'+review, allow_pickle=True)
    df = pd.DataFrame(rev, columns=['created_at', 'writer_id', 'praise_count',
                               'prod_id', 'prod_name', 'prod_opt',
                               'comment', 'image_url'])
    
    df.drop_duplicates(['comment'],keep = 'first',inplace=True)
    df.reset_index(drop=True, inplace=True)

    # text cleaning
    com = df['comment'].to_numpy()
    com_clean = text_cleaning(com)

    # spell checking
    com_spellcheck = hanspell_(com_clean)
    com = com_spellcheck

    # 내보내기
    df['comment_spellcheck'] = com
    df.to_csv(save_path+review.split('.')[0]+'_spellcheck.csv', index=False)
