==> Model/3)Mecab_QnA.ipynb <==
# Mecab QnA

## 라이브러리

#pip freeze

import os

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

from tqdm.notebook import tqdm

import re
import pickle

from konlpy.tag import Mecab

## Path

base_path = '../Data/QnA/'
save_path = '../Data/QnA/'

## Load Data

df = pd.read_csv(base_path+'qna_table_hanspell.csv')

df.shape

# hanspell 이후 <None> 제거
df = df[(df['question_spellcheck']!='<None>')&(df['answer_spellcheck']!='<None>')]
df.reset_index(drop=True, inplace=True)

print("<None> 제거 후 QnA 갯수: " + str(df.shape))

print("label별 QnA 갯수 확인")
df['label'].value_counts()

## Mecab

mecab = Mecab()

def make_mecab_tokens(comment):

    # mecab 돌리기
    comment_mecab = []
    for i in range(len(comment)):
        if comment[i] == 'NaN': #NaN값 처리ㅣ
            comment_mecab.append('nan')
        else:
            comment_mecab.append(mecab.pos(comment[i]))

    # 조사 제거
    morpheme = ['NNG','NNP','NNB','NNBC','NR','NP','VV','VA','VX','VCP','VCN','MM','MAG','MAJ','IC','SN']    
    tmp = []
    comment_tokens = []
    
    for tok in comment_mecab:
        for t in tok:
            if t[1] in morpheme:
                tmp.append(t[0])
        if len(tmp)==0:
            comment_tokens.append('')
        else:
            comment_tokens.append(tmp)
        tmp = []

    # 한 문장으로 결합
    comment_tokens_str = [' '.join(re) for re in comment_tokens]

    return comment_tokens_str

# question
question = df['question_spellcheck'].astype(str)

df['question_mecab_SN'] = make_mecab_tokens(question)

# answer
answer = df['answer_spellcheck'].astype(str)

df['answer_mecab_SN'] = make_mecab_tokens(answer)

df = df[(df['question_mecab_SN']!='')&(df['answer_mecab_SN']!='')] # NaN 제거
df.reset_index(drop=True, inplace=True)

df.shape

df[:1]

## Labeling

# label 수치형으로 변환
cat = pd.Series(df['label'])
labels, uniques = pd.factorize(cat)
labels, uniques

df['num_label'] = labels
df['num_label'] = df['num_label'].astype(str) # str으로 변환

df[['label','num_label']].value_counts()

#기타와 반품 label 순서 변경
df.loc[df['num_label'] =='4', 'num_label'] = 'tmp'
df.loc[df['num_label'] =='5', 'num_label'] = '4'
df.loc[df['num_label'] =='tmp', 'num_label'] = '5'

df[['label','num_label']].value_counts()

df.shape

# csv 저장 후 내보내기
df.to_csv(save_path+'qna_table_mecab_6_SN.csv', index=False)

## KoBERT용 데이터 생성

* 기타(5) label 제거
* Que + '\t' + Ans + '\t' 형태
* txt 파일로 내보내기

# 기타 제거
df_drop = df[df['label']!='기타']

# Que + '\t' + Ans + '\t'
data_drop = pd.DataFrame()
data_drop['BERT_NEW'] = df_drop['question_mecab_SN'].values + '\t' + df_drop['answer_mecab_SN'].values + '\t' + df_drop['num_label'].astype(str).values + '\n'

data_kb = data_drop['BERT_NEW']
data_train, data_test = train_test_split(data_kb, test_size=0.2, shuffle=True, random_state=316)

# 내보내기
with open(save_path+'type1_SN_BERT_drop_test_new.txt', 'w') as file:
    file.writelines(data_test.values.tolist())
    
with open(save_path+'type1_SN_BERT_drop_train_new.txt', 'w') as file:
    file.writelines(data_train.values.tolist())
