==> Model/3)Mecab_review.ipynb <==
# Review

## Library

import pandas as pd
import numpy as np
import os
from tqdm import tqdm, tqdm_notebook

from konlpy.tag import Mecab

## Path

open_path = '../Data/Review/Review_embedding/'
save_path = '../Data/Review/Review_mecab_SN/'

## Mecab 돌려서 csv로 저장

mecab = Mecab()

review_list = os.listdir(open_path)

def make_review_mecab_tokens(file_name):
    review_df = pd.read_csv(open_path+file_name,engine='python')
    review_df.drop(columns = ['comment_mecab','review_embedding','review_class'],inplace=True)
    
    review = review_df['comment_spellcheck'].astype(str)    

    # mecab 돌리기
    review_mecab = []
    for i in range(len(review)):
        if review[i] == 'NaN':
            review_mecab.append('nan')
        else:
            review_mecab.append(mecab.pos(review[i]))

    # 조사 제거
    morpheme = ['NNG','NNP','NNB','NNBC','NR','NP','VV','VA','VX','VCP','VCN','MM','MAG','MAJ','IC','SN']    
    tmp = []
    review_tokens = []
    
    for tok in review_mecab:
        for t in tok:
            if t[1] in morpheme:
                tmp.append(t[0])
        if len(tmp)==0:
            review_tokens.append('')
        else:
            review_tokens.append(tmp)
        tmp = []

    # 한 문장으로 결합
    review_tokens_str = [' '.join(re)+'\n' for re in review_tokens]
    review_df['comment_mecab'] = review_tokens_str

    review_df = review_df[review_df['comment_mecab']!='\n'] # NaN 제거

    # csv 저장 후 내보내기    
    review_name = file_name.split('embedding')[0] + 'mecab.csv'
    review_df.to_csv(save_path+review_name, index=False)

# 전체 리뷰 파일 함수 돌리기
for file_name in tqdm_notebook(review_list):
    make_review_mecab_tokens(file_name)
