==> Model/5)Embedding_Review.ipynb <==
# SKT-KoBERT

https://github.com/SKTBrain/KoBERT

## Install & Library

import torch
from torch import nn
import torch.nn.functional as F
import torch.optim as optim
from torch.utils.data import Dataset, DataLoader
import gluonnlp as nlp
import pandas as pd
import numpy as np
from tqdm import tqdm, tqdm_notebook
import os
import pickle

from KoBERT.kobert.utils import get_tokenizer
from KoBERT.kobert.pytorch_kobert import get_pytorch_kobert_model

from transformers import AdamW
from transformers.optimization import get_cosine_schedule_with_warmup

np.set_printoptions(formatter={'float_kind': lambda x: "{0:0.5f}".format(x)})

# CPU or GPU
device = 'cuda:0' if torch.cuda.is_available() else 'cpu'
device

## Path

model_path = './Model/'
open_path = '../Data/Review/Review_mecab_SN/'
save_path = '../Data/Review/Review_embedding_SN_new/'
pickle_path = '../Data/Review/'

## Model & Tokenizer

bertmodel, vocab = get_pytorch_kobert_model()

tokenizer = get_tokenizer()
tok = nlp.data.BERTSPTokenizer(tokenizer, vocab, lower=False)

class BERTDataset(Dataset):
    def __init__(self, dataset, sent_idx, bert_tokenizer, max_len,
                 pad, pair):
        transform = nlp.data.BERTSentenceTransform(
            bert_tokenizer, max_seq_length=max_len, pad=pad, pair=pair)

        self.sentences = [transform([i]) for i in dataset] # sent_idx를 제거

    def __getitem__(self, i):
        return (self.sentences[i])

    def __len__(self):
        return (len(self.sentences))

## Setting parameters
max_len = 210 #수정
batch_size = 16 #수정
warmup_ratio = 0.1
num_epochs = 5
max_grad_norm = 1
log_interval = 200
learning_rate =  5e-5

class BERTClassifier(nn.Module):
    def __init__(self,
                 bert,
                 hidden_size = 768,
                 num_classes=5, # label 갯수에 맞춰 수정
                 dr_rate=None,
                 params=None):
        super(BERTClassifier, self).__init__()
        self.bert = bert
        self.dr_rate = dr_rate
                 
        self.classifier = nn.Linear(hidden_size , num_classes)
        if dr_rate:
            self.dropout = nn.Dropout(p=dr_rate)
    
    def gen_attention_mask(self, token_ids, valid_length):
        attention_mask = torch.zeros_like(token_ids)
        for i, v in enumerate(valid_length):
            attention_mask[i][:v] = 1
        return attention_mask.float()

    def forward(self, token_ids, valid_length, segment_ids):
        attention_mask = self.gen_attention_mask(token_ids, valid_length)
        
        _, pooler = self.bert(input_ids = token_ids, token_type_ids = segment_ids.long(), attention_mask = attention_mask.float().to(token_ids.device))
        
        classifier = self.classifier(pooler)
        
        if self.dr_rate:
            out = self.dropout(classifier)
  
        return pooler, out

model = BERTClassifier(bertmodel, dr_rate=0.5).to(device)

model.load_state_dict(torch.load(model_path+'type1_KoBERT_new.pt')) # 모델 불러오기

model.eval()

def review_BERTClassifier(_dataloader):
    
    # 저장할 빈 리스트 생성  
    review_embedding = []
    review_probability = []

    for batch_id, (token_ids, valid_length, segment_ids) in enumerate(_dataloader):

        token_ids = token_ids.long().to(device)
        segment_ids = segment_ids.long().to(device)
        valid_length= valid_length  

        embedding, probability = model.forward(token_ids, valid_length, segment_ids)

        embedding = embedding.cpu().detach().numpy()
        probability = probability.cpu().detach().numpy()

        # list에 append
        review_embedding.append(embedding)
        review_probability.append(probability)

        del embedding, probability
        torch.cuda.empty_cache() # GPU 캐시 삭제
    
    # concat
    review_embedding_ = np.concatenate([rev for rev in review_embedding], axis=0)
    review_probability_ = np.concatenate([rev for rev in review_probability], axis=0)
    
    # max index 추출
    review_class = np.argmax(np.array(review_probability_), axis=1)
    
    return review_embedding_, review_class

## Review Classification

review_list = os.listdir(open_path)
review_name = [review.split('_mecab')[0] for review in review_list]

len(review_list) #.ipynb_checkpoints 파일이 존재하여 146, review는 총 145

for i in tqdm_notebook(range(len(review_list))): 
    if review_list[i] == '.ipynb_checkpoints':
        continue
    
    # load data
    df = pd.read_csv(open_path + review_list[i])
    df=df[df['comment_mecab'].notna()]
    df.reset_index(drop=True, inplace=False)
    
    tokens = df['comment_mecab'].tolist()
    
    data = BERTDataset(tokens, 0, tok, max_len, True, False)
    _dataloader = torch.utils.data.DataLoader(data, batch_size=batch_size, num_workers=5)
    
    # classifier
    review_embedding, review_class = review_BERTClassifier(_dataloader)
    
    # length 체크
    if len(df)!=len(review_embedding.tolist()):
        print('not match length:', i)
    
    # csv 파일로 내보내기
    df['review_embedding'] = review_embedding.tolist()
    df['review_class'] = review_class.tolist()
    
    # pickle 파일로 저장
    with open(save_path+review_name[i]+'_embedding.pkl', "wb") as file:
        pickle.dump(df, file)
    
    # df.to_csv(save_path+review_name[i]+'_embedding'+'.csv', index=False)

# 제대로 저장되었는지 확인
make_list = os.listdir(save_path)

len(make_list)

---
### 상품별 리뷰 개수 확인 & 전체 리뷰에 대한 label 개수 표 생성

df = pd.DataFrame(columns=['file_name', 'length', '상품_0', '배송_1', '교환_2', '환불_3', '반품_4'])

for names in tqdm_notebook(make_list):
    tmp = pd.read_pickle(save_path+names)
    app_list = [names, len(tmp), len(tmp[tmp['review_class']==0]), len(tmp[tmp['review_class']==1]), 
                len(tmp[tmp['review_class']==2]), len(tmp[tmp['review_class']==3]), len(tmp[tmp['review_class']==4])]
    df.loc[len(df)] = app_list  

### 전체 리뷰에 대한 label 개수 표 생성
pd.DataFrame(df.iloc[:,1:].sum(), columns=['count'])

### 리뷰가 가장 많은 2개 데이터 선정
df.sort_values(by='length', ascending=False)
# >> review_344496_embedding.csv & review_388715_embedding.csv

df['length'].median() #리뷰 갯수의 중앙값 확인

## Save Data

pickle_path = '/home/users/hyejeongeun/Todays_Chatbot/Data/Review/'

with open(pickle_path+"review_embedding_SN_new.pkl", "wb") as file:
    pickle.dump(df, file)
