==> Model/6)Matching_System.ipynb <==
# QnA System

## Install & Library

# !nvidia-smi

import numpy as np
import pandas as pd
import pickle
from tqdm import tqdm, tqdm_notebook

from KoBERT.kobert.utils import get_tokenizer
from KoBERT.kobert.pytorch_kobert import get_pytorch_kobert_model

from konlpy.tag import Mecab

from sklearn.metrics.pairwise import cosine_similarity

import torch
from torch import nn
import torch.nn.functional as F
import torch.optim as optim
from torch.utils.data import Dataset, DataLoader
import gluonnlp as nlp

from transformers import AdamW
from transformers.optimization import get_cosine_schedule_with_warmup

import warnings
warnings.filterwarnings(action='ignore')

pd.set_option('max_row', None)
pd.set_option('display.max_colwidth', -1)

np.set_printoptions(formatter={'float_kind': lambda x: "{0:0.5f}".format(x)})

device = 'cuda:1' if torch.cuda.is_available() else 'cpu'
device

## Path

model_path = './Model/'
qna_path = '../Data/QnA/'
review_path = '../Data/Review/Review_embedding_SN_new/'
data_path = '../Data/'

## BERT Model

class BERTDataset(Dataset):
    def __init__(self, dataset, sent_idx, bert_tokenizer, max_len,
                 pad, pair):
        transform = nlp.data.BERTSentenceTransform(
            bert_tokenizer, max_seq_length=max_len, pad=pad, pair=pair)

        self.sentences = [transform([i]) for i in dataset] # sent_idx를 제거
        
    def __getitem__(self, i):
        return (self.sentences[i])

    def __len__(self):
        return (len(self.sentences))

class BERTClassifier(nn.Module):
    def __init__(self,
                 bert,
                 hidden_size = 768,
                 num_classes=5,  ##### 수정
                 dr_rate=None,
                 params=None):
        super(BERTClassifier, self).__init__()
        self.bert = bert
        self.dr_rate = dr_rate
                 
        self.classifier = nn.Linear(hidden_size , num_classes)
        if dr_rate:
            self.dropout = nn.Dropout(p=dr_rate)
    
    def gen_attention_mask(self, token_ids, valid_length):
        attention_mask = torch.zeros_like(token_ids)
        for i, v in enumerate(valid_length):
            attention_mask[i][:v] = 1
        return attention_mask.float()

    def forward(self, token_ids, valid_length, segment_ids):
        attention_mask = self.gen_attention_mask(token_ids, valid_length)
        
        _, pooler = self.bert(input_ids = token_ids, token_type_ids = segment_ids.long(), attention_mask = attention_mask.float().to(token_ids.device))
        
        classifier = self.classifier(pooler)
        
        if self.dr_rate:
            out = self.dropout(classifier)
  
        return pooler, out

## Parameter

## Setting parameters
max_len = 210 #수정
batch_size = 16 # 64 
warmup_ratio = 0.1
num_epochs = 5
max_grad_norm = 1
log_interval = 200
learning_rate =  5e-5

# load model
bertmodel, vocab = get_pytorch_kobert_model()
tokenizer = get_tokenizer()
tok = nlp.data.BERTSPTokenizer(tokenizer, vocab, lower=False)

# load pre-trained KoBERTClassifier
model = BERTClassifier(bertmodel, dr_rate=0.5).to(device)
model.load_state_dict(torch.load(model_path+'type1_KoBERT_new.pt'))
model.eval() 

## Mecab

# load mecab
mecab = Mecab()

def make_question_mecab_tokens(question):
    # mecab 돌리기
    que_mecab = mecab.pos(question[0])

    # 조사 제거
    morpheme = ['NNG','NNP','NNB','NNBC','NR','NP','VV','VA','VX','VCP','VCN','MM','MAG','MAJ','IC','SN'] # SN 추가
    tmp = []
    que_tokens = []

    for t in que_mecab:
        if t[1] in morpheme:
            que_tokens.append(t[0])

    if len(que_tokens)==0:
        que_tokens.append('')

    # 한 문장으로 결합
    que_tokens_str = [' '.join(que_tokens)]
    
    return que_tokens_str

## Input Embedding & Classification

def input_BERTClassifier(model, test_dataloader):
    for batch_id, (token_ids, valid_length, segment_ids) in enumerate(test_dataloader):

        token_ids = token_ids.long().to(device)
        segment_ids = segment_ids.long().to(device)
        valid_length= valid_length  
        
        test_embedding, test_probability = model.forward(token_ids, valid_length, segment_ids)

        test_embedding = np.array(test_embedding.tolist())
        test_probability = np.array(test_probability.tolist())

        torch.cuda.empty_cache() # GPU 캐시 삭제
        
    # class 
    test_class = np.argmax(np.array(test_probability))
        
    return test_embedding, test_class

## Semantic Search

def Semantic_Search_Question(question, input_embedding, input_class, product_id):
    # load file
    df = pd.read_pickle(qna_path+'question_embedding_SN_new.pkl')
    
    # 같은 상품 내의 질문만 추출
    df = df[df['product_id'] == product_id]
    df.reset_index(drop=True, inplace=True)

    # input_class와 같은 label만 추출
    label_idx = np.array(np.where(np.array(df['question_class'].tolist()) == input_class)).tolist()[0]
    df_in_input = df.loc[label_idx]
    df_in_input.reset_index(drop=True, inplace=True)
    
    # 질문 embedding
    question_embedding = df_in_input['question_embedding']

    # make tensor
    corpus_embeddings = torch.tensor(question_embedding)
    query_embedding = torch.tensor(input_embedding)
    
    # cosine similarity 계산
    cos_scores = cosine_similarity(query_embedding, corpus_embeddings)
    cos_socres_round = np.array(list(map(lambda x:int(x*1000)/1000, cos_scores[0]))) 
    df_in_input['input_question'] = np.repeat(question, len(df_in_input))
    df_in_input['score'] = cos_socres_round
    
    # 0.95 이상만 반환
    df_in_input_top = df_in_input[df_in_input['score']>=0.95]
    df_in_input_top.sort_values(by='score', ascending=False, inplace=True)
    df_in_input_top.reset_index(drop=True, inplace=True)
    
    return df_in_input_top

def Semantic_Search_Review(question, input_embedding, input_class, product_id):
    # load review data
    df = pd.read_pickle(review_path+'review_'+str(product_id)+'_embedding.pkl')
    
    # input_class와 같은 label만 추출
    label_idx = np.array(np.where(np.array(df['review_class'].tolist()) == input_class)).tolist()[0]
    
    if len(label_idx) != 0: # 같은 label의 리뷰가 하나라도 있으면
        df_in_input = df.loc[label_idx]
        df_in_input['using_label'] = input_class
        df_in_input.reset_index(drop=True, inplace=True)    
    else: # 하나도 없으면
        return pd.DataFrame() # 빈 데이터 프레임 반환
    
    # 리뷰 embedding
    review_embedding = df_in_input['review_embedding']
              
    # make tensor
    corpus_embeddings = torch.tensor(review_embedding)
    query_embedding = torch.tensor(input_embedding)
    
    # cosine similarity 계산
    cos_scores = cosine_similarity(query_embedding, corpus_embeddings)
    cos_socres_round = np.array(list(map(lambda x:int(x*1000)/1000, cos_scores[0]))) 
    df_in_input.loc[:,'input_question'] = np.repeat(question, len(df_in_input))
    df_in_input['score'] = cos_socres_round
    
    # 0.95 이상만
    df_in_input_top = df_in_input[df_in_input['score']>=0.95]
    df_in_input_top.sort_values(by='score', ascending=False, inplace=True)
    df_in_input_top.reset_index(drop=True, inplace=True)
    
    return df_in_input_top

## Custom

### Sentiment Analysis

def Text_Sentiment_Review(df):
    # sentiment table 불러오기
    with open(data_path+'table.pickle', 'rb') as file:
        table = pickle.load(file)
 
    # 초기화
    df['neg'] = 0
    df['neut'] = 0
    df['pos'] = 0

    # df 리뷰 전체 리스트
    review_list = df['comment_mecab'].apply(lambda x:x.replace('\n','').split(' ')).values.tolist()

    # table에 있는 경우만 남기기
    review_list = [[word for word in review if word in table] for review in review_list]

    # 리스트 내 word를 table 내 숫자로 바꾸기
    neg_list = [[float(table[word]['Neg']) for word in review] for review in review_list]
    pos_list = [[float(table[word]['Pos']) for word in review] for review in review_list]

    # 값 sum()
    neg = [sum(c) for c in neg_list]
    pos = [sum(c) for c in pos_list]

    # df 문장 단어 개수
    df['count'] = df['comment_mecab'].apply(lambda x:len(x.replace('\n','').split(' ')))

    # df에 구한 값 추가 & 개수로 나누기
    df['neg'] = round(neg/df['count'], 3)
    df['pos'] = round(pos/df['count'], 3)

    # 부정 점수 0.3 이하 & 긍정 점수 0.3 이상인 경우만 추출
    df = df[(df['neg']<=0.3)&(df['pos']>=0.3)]
    
    # column drop
    df_return = df.drop(columns=['count'])

    return df_return

### Word Count

def word_count_in_wordcloud_Review(df):
    # wordcloud dict 불러오기   
    with open(data_path+'word_dict.pickle', 'rb') as file:
        word_dict = pickle.load(file)
    
    # 전체 단어 count 개수
    df['word_count'] = df['comment_mecab'].apply(lambda x: len(x.split(' ')))
    
    # 예측된 label의 wordcloud에 포함된 단어
    df['word_count_in_cloud'] = 0
    for i in range(len(df)):
        df['word_count_in_cloud'][i] = len([w for w in df['comment_mecab'][i].replace('\n','').split(' ') if w in(word_dict[df['review_class'][i]][:30])]) # 30개만 불러오기
    
    # column 제거
    df_return = df.drop(columns=['word_count_in_cloud'])
    
    return df_return

## Select Top

def select_top2_Question(df):
    # score 기준 내림차순으로 정렬
    df.sort_values(by=['score'], ascending=False, inplace=True)
    df.reset_index(drop=True, inplace=True) # reset index
        
    if df.shape[0] < 2: # 개수가 2개보다 적으면
        df_return = df
    else:
        df_return = df[:2]

    # 필요한 column만 반환
    df_return = df_return[['input_question', 'score', 'question', 'answer']]

    return df_return

def select_top5_Review(df, sentiment, wordcount):    
    # Custom 여부에 따라 정렬
    if (sentiment == False) and (wordcount == False):
        sort_list = ['score', 'praise_count', 'image']
        df.sort_values(by=sort_list, ascending=False, inplace=True)
        
    if (sentiment == True) and (wordcount == False):
        sort_list = ['score', 'pos', 'neg', 'praise_count', 'image']
        df.sort_values(by=sort_list, ascending=[False, False, True, False, False], inplace=True)
    
    if (sentiment == False) and (wordcount == True):
        sort_list = ['score', 'word_count', 'praise_count', 'image']
        df.sort_values(by=sort_list, ascending=False, inplace=True)
    
    if (sentiment == True) and (wordcount == True):
        sort_list = ['score', 'pos', 'neg', 'word_count', 'praise_count', 'image']
        df.sort_values(by=sort_list, ascending=[False, False, True, False, False, False], inplace=True)  
        
    # reset index
    df.reset_index(drop=True, inplace=True)
    
    if df.shape[0] < 5: # 5개보다 적으면
        df_return = df
    else:
        df_return = df[:5]
    
    # 필요한 column만 추출
    out_list = ['prod_id', 'prod_name', 'input_question', 'comment'] + sort_list
    df_return = df_return[out_list]
    
    return df_return

## Matching System

def matching_system(product_id, question, sentiment, wordcount):
    # run mecab
    que_mecab = make_question_mecab_tokens(question)

    # make dataloader
    input_data = BERTDataset(que_mecab, 0, tok, max_len, True, False)
    input_dataloader = torch.utils.data.DataLoader(input_data, batch_size=batch_size, num_workers=5)
    
    # input data embedding & class
    input_embedding, input_class = input_BERTClassifier(model, input_dataloader)
    
    
    # Semantic Search between input & question
    question_top = Semantic_Search_Question(question, input_embedding, input_class, product_id)
    # select qna top2
    question_top2 = select_top2_Question(question_top)
    
    
    # Semantic Search between input & review
    review_top = Semantic_Search_Review(question, input_embedding, input_class, product_id)
    # Make sentiment column
    if sentiment == True: 
        review_top = Text_Sentiment_Review(review_top)
    # Make word count column
    if wordcount == True:
        review_top = word_count_in_wordcloud_Review(review_top)
    # select review top5
    review_top5 = select_top5_Review(review_top, sentiment, wordcount)  

    return question_top2, review_top5

----

## Run Matching System

# custom 모델 list
tf_list = [['Base', False, False], ['Sentiment', True, False], ['Wordcount', False, True]]

# 매칭 시스템 실행 함수
def evaluation(id_, question_):
    df_display = pd.DataFrame(columns=['input_question'])
    for tf in tf_list:
        question, review = matching_system(id_, question_, 
                                           tf[1], # sentiment
                                           tf[2]) # wordcount      
        
        if tf[0] == 'Base':
            df_display = pd.concat([df_display, question[['input_question','question','answer']]])
            display(question[['input_question','question','answer']])
        print('이번 결과는: ', tf[0])
        df_display = pd.concat([df_display, review[['input_question','comment']]])
        display(review[['input_question','comment']])
    return df_display

# 예시
id_ = 388715
question_ = '나사 조립이 많이 어려운가요'

evaluation(id_, question_)
